+++ 05_fifthApp/package.json
@@ -15,6 +15,8 @@
     "react-hook-form": "^7.43.9",
     "react-redux": "^8.0.5",
     "react-scripts": "5.0.1",
+    "redux-thunk": "^2.4.2",
+    "reselect": "^4.1.8",
     "sass": "^1.62.1",
     "uuid": "^9.0.0",
     "web-vitals": "^2.1.4"
