+++ 05_fifthApp/src/reducers/index.js
@@ -1,77 +1,2 @@
-const initialState = {
-  heroes: [],
-  heroesLoadingStatus: "idle",
-  heroCreatingStatus: "idle",
-  heroRemovingStatus: "idle",
-  filters: [],
-  activeFilters: ["all"],
-};
-
-export const reducer = (state = initialState, action) => {
-  switch (action.type) {
-    case "ELEMENT_OPTIONS_FETCHED":
-      return {
-        ...state,
-        filters: action.payload,
-      };
-    case "HEROES_FETCHING":
-      return {
-        ...state,
-        heroesLoadingStatus: "loading",
-      };
-    case "HEROES_FETCHED":
-      return {
-        ...state,
-        heroes: action.payload,
-        heroesLoadingStatus: "idle",
-      };
-    case "HEROES_FETCHING_ERROR":
-      return {
-        ...state,
-        heroesLoadingStatus: "error",
-      };
-
-    case "HERO_CREATING":
-      return {
-        ...state,
-        heroCreatingStatus: "loading",
-      };
-    case "HERO_CREATED":
-      return {
-        ...state,
-        heroes: [...state.heroes, action.payload],
-        heroCreatingStatus: "idle",
-      };
-    case "HERO_CREATING_ERROR":
-      return {
-        ...state,
-        heroCreatingStatus: "error",
-      };
-
-    case "HERO_REMOVING":
-      return {
-        ...state,
-        heroRemovingStatus: "loading",
-      };
-    case "HERO_REMOVED":
-      return {
-        ...state,
-        heroes: [...action.payload],
-        heroRemovingStatus: "idle",
-      };
-    case "HERO_REMOVING_ERROR":
-      return {
-        ...state,
-        heroRemovingStatus: "error",
-      };
-
-    case "TOGGLE_FILTER":
-      return {
-        ...state,
-        activeFilters: action.payload,
-      };
-
-    default:
-      return state;
-  }
-};
+import { filters } from "./filters";
+import { heroes } from "./heroes";
